Add files via upload
This project is the implementation of bitcoin concepts. It generates target with the number of 0's you specify with the difficulty. Then it generates solution which satisfies the condition hash(m||s)<=t and then verifies the solution which is generated . 
Also one function (differentsolution() ) is taking out 6 different solutions for 21=<I<=26 and calculating time.